requirements.txt
google-cloud-vision==2.6.3
streamlit
Pillow
cv2